Add files via upload
Title: "Stock Price Forecasting and Regression Analysis: Apple vs. Honeywell"

Description: This project focuses on forecasting the future share prices of two prominent companies, Apple and Honeywell, using various forecasting methods and regression analysis. The assignment begins with an exploration of forecasting methods followed by the application of regression models to predict stock prices. Additionally, residual analysis of the simple regression model is performed to assess the validity of the regression approach. The conclusion of the analysis reveals that while regression analysis produced the most accurate forecasts compared to other methods, residual analysis uncovered potential issues such as autocorrelation and heteroscedasticity, particularly in the case of Apple. Despite these challenges, regression analysis remains a valuable tool for stock price forecasting, provided that the assumptions of independence and homoscedasticity are carefully evaluated and maintained.